+++ helpmehelpyou/src/Components/GiveHelp/giveHelp.js
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import "../../giveAndFindHelp.css";
+import "../giveAndFindHelp.css";
 import image from "./givehelp-image.png";
 import dummyData from "./DummyData";
 import { Link } from "react-router-dom";
@@ -20,8 +20,6 @@ export default function GiveHelp() {
     setUserInput(event.target.value);
     console.log(userInput);
   }
-
- 
 
   function contactUser(email) {
     // Display email address of user
@@ -48,32 +46,46 @@ Listing - <h1> for title / summary
       ></img>
       {/* Search bar and button */}
       <section id="give-and-find-help-search-section">
-        <input type="text" onChange={storeInput} className="give-and-find-help-search-box"></input>
+        <input
+          type="text"
+          onChange={storeInput}
+          className="give-and-find-help-search-box"
+        ></input>
         <button className="give-and-find-help-search-button">🔍</button>
-      </section>  <section id="give-help-post-request-link-from-givehelp">      
-
-        <p className="give-help-post-request">Need help? <Link to ="/findhelpform">Post a request </Link></p>
-
+      </section>{" "}
+      <section id="give-help-post-request-link-from-givehelp">
+        <p className="give-help-post-request">
+          Need help? <Link to="/findhelpform">Post a request </Link>
+        </p>
       </section>
       {/* Recent listings fetched from DB */}
       <section id="give-help-listings">
-        <h3 className="give-and-find-help-listings-area-title">Recent listings</h3>
+        <h3 className="give-and-find-help-listings-area-title">
+          Recent listings
+        </h3>
         {dummyData.map((listing) => (
-          <div key={listing.listing_id} className="give-and-find-help-individual-listing">
-            <h1 className="give-and-find-help-listing-title">{listing.listing_title}</h1>
+          <div
+            key={listing.listing_id}
+            className="give-and-find-help-individual-listing"
+          >
+            <h1 className="give-and-find-help-listing-title">
+              {listing.listing_title}
+            </h1>
             <div className="give-help-sub-title">
               {/* Are there skills wanted in the guest listings DB? */}
               <p className="give-and-find-help-skills">{listing.ssc_wanted}</p>
               {/* Date stamp - this needs editing */}
               <p className="give-help-date-posted">{listing.created_at}</p>
             </div>
-            <p className="give-help-listing-details">{listing.listing_details}</p>
+            <p className="give-help-listing-details">
+              {listing.listing_details}
+            </p>
             <p className="give-help-skills-offer-details">
               I can offer... {listing.skills_offered}
             </p>
             {/* This div isn't being used at the moment and was messing up the alignment of the box! When an image is added - this div can be added again. 
             <div className="give-help-user-info"> */}
-              {/* <div className="give-help-image-container">
+            {/* <div className="give-help-image-container">
                 {/* There is no image in the DB at the moment 
                 <img
                   className="give-help-profile-picture"
@@ -83,17 +95,21 @@ Listing - <h1> for title / summary
                 There is no rating yet - we need to do a join to the user table 
                 <p className="give-help-rating">{listing.rating}</p>
               </div> */}
-              <div className="give-help-user-details-container">
-                <div className="give-help-posted-by-div">
+            <div className="give-help-user-details-container">
+              <div className="give-help-posted-by-div">
                 <p className="give-and-find-help-display-name">Posted by:</p>
-                <p className="give-and-find-help-display-name">{listing.display_name}</p>
-                </div>
-                <div className="give-help-location-div">
-                <p className="give-and-find-help-display-name">Location:</p>
-                <p className="give-and-find-help-display-name">{listing.borough_name}</p>
-                </div>
-                <p className="give-help-borough-name"> </p>
+                <p className="give-and-find-help-display-name">
+                  {listing.display_name}
+                </p>
               </div>
+              <div className="give-help-location-div">
+                <p className="give-and-find-help-display-name">Location:</p>
+                <p className="give-and-find-help-display-name">
+                  {listing.borough_name}
+                </p>
+              </div>
+              <p className="give-help-borough-name"> </p>
+            </div>
             {/* </div> */}
             <div className="give-and-find-help-user-contact">
               <button
@@ -102,7 +118,9 @@ Listing - <h1> for title / summary
               >
                 Contact user
               </button>
-              <button className="give-and-find-help-visit-profile">View Profile</button>
+              <button className="give-and-find-help-visit-profile">
+                View Profile
+              </button>
             </div>
           </div>
         ))}
